% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.utils.R
\name{cal.signatureScore.gdT.Fred}
\alias{cal.signatureScore.gdT.Fred}
\title{Identification of gamma delta T cells (Fred's method)}
\usage{
cal.signatureScore.gdT.Fred(
  obj,
  GSx = c("CD3D", "CD3E", "TRDC", "TRGC1", "TRGC2"),
  GSy = c("CD8A", "CD8B"),
  col.name = "Score.gammaDeltaT",
  th.score = 0.35,
  out.prefix = NULL
)
}
\arguments{
\item{obj}{object of \code{SingleCellExperiment}; assay "counts" is required.}

\item{GSx}{character vector; gdT marker gene list. (default: c("CD3D","CD3E","TRDC","TRGC1","TRGC2")).}

\item{GSy}{character vector; CD8 T cell marker gene list. (default: c("CD8A","CD8B")).}

\item{col.name}{character; prefix of column names, to be added to obj. (default: "Score.gammaDeltaT")}

\item{th.score}{double; threshold of the score. (default: 0.35)}

\item{out.prefix}{character; output prefix. (default: NULL)}
}
\value{
a SingleCellExperiment object
}
\description{
Identification of gamma delta T cells (Fred's method)
}
