% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{calEffectSizeFromDE}
\alias{calEffectSizeFromDE}
\title{calculate effect size given a de.out object}
\usage{
calEffectSizeFromDE(de.out, de.mode = "multiAsTwo", ncores = 8, cal.p = F)
}
\arguments{
\item{de.out}{list; Three components are required: two data.table named "all" and "sig", a list named "fit"}

\item{de.mode}{character; mode of differential expression analysis. (default: "multiAsTwo")}

\item{ncores}{integer; number of CPU cores to use. (default: 8)}

\item{cal.p}{logical; whether to calculate p-values assuming normal distribution on the z-scores (default: F)}
}
\value{
a list contain components, "es.tb" (data.table), "ncells.vec" and "ncells.control.vec" (integer vectors)
}
\description{
calculate effect size given a de.out object
}
\details{
calculate effect size given a de.out object
}
