% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{convertLimmaToSCE}
\alias{convertLimmaToSCE}
\title{convert the limma result to gene by meta-cluster data stored in an SingleCellExperiment object}
\usage{
convertLimmaToSCE(
  de.limma.tb,
  out.prefix,
  ncores = 8,
  min.ncells = 30,
  min.ncellsStudy = 200,
  gset.list = NULL,
  direct.sig = T,
  de.mode = "multiAsTwo",
  column.exp = "meanScale",
  gene.used = NULL,
  colSet = list()
)
}
\arguments{
\item{de.limma.tb}{data.table; one line for a dataset. columns "data.id", "platform", and "dfile" are required}

\item{out.prefix}{character; output prefix}

\item{ncores}{integer; number of CPU cores to use. (default: 8)}

\item{min.ncells}{integer; only meta-clusters with number of cells > min.ncells are used. (default: 30)}

\item{min.ncellsStudy}{integer; only datasets with number of cells > min.ncellsStudy are used. (default: 200)}

\item{gset.list}{list; list containing gene sets. (default: NULL)}

\item{direct.sig}{logical; if TRUE, genes exibit significance in all datasets will be assigned "sig" directly, irrespective of combined ES and combined adjusted p vlaue. (default: TRUE)}

\item{de.mode}{character; mode of differential expression analysis. (default: "multiAsTwo")}

\item{column.exp}{character; convert the column of limma result to assay data. (default: "meanScale")}

\item{gene.used}{character; only keep genes in gene.used. (default: NULL)}

\item{colSet}{list; mapping iterms in the names to colors in the values. (default: list())}
}
\value{
a SingleCellExperiment object
}
\description{
convert the limma result to gene by meta-cluster data stored in an SingleCellExperiment object
}
\details{
convert the limma result to gene by meta-cluster data stored in an SingleCellExperiment object
}
