% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.utils.R
\name{fill.contamination}
\alias{fill.contamination}
\title{calculate the signature score of one potential contaminated cell types (simple average-threshold method)}
\usage{
fill.contamination(
  obj,
  out.prefix,
  assay.name = "norm_exprs",
  g.name = "plasmaB",
  g.test = c("CD79A", "JCHAIN", "SDC1"),
  score.t = 1,
  vis.v = c(0.25, 0.5, 1)
)
}
\arguments{
\item{obj}{object of \code{SingleCellExperiment} or \code{Seurat}}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{assay.name}{character vector; which assay to use. (default: "norm_exprs").}

\item{g.name}{character; signature name of the cell type. (default: "plasmaB").}

\item{g.test}{character vector; signature genes of the cell type. (default: c("CD79A", "JCHAIN", "SDC1")).}

\item{score.t}{double; threshold for signature score (mean of g.test). (default: 1)}

\item{vis.v}{double vector; for vertical lines in visulization. (default: c(0.25,0.5,1)).}
}
\value{
an object of \code{SingleCellExperiment} or \code{Seurat}
}
\description{
calculate the signature score of one potential contaminated cell types (simple average-threshold method)
}
\details{
Two columns will be added to the obj, "%s.score" and "%s.class, where %s is the g.name.
}
