% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{HVG.From.GeneRankTb}
\alias{HVG.From.GeneRankTb}
\title{highly variable genes finding (method using percetile rank values from F test)}
\usage{
HVG.From.GeneRankTb(
  gene.rank.tb,
  n.common = 1000,
  n.specific = 1000,
  th.rank = 0.1
)
}
\arguments{
\item{gene.rank.tb}{data.table object}

\item{n.common}{integer; number of common top genes. (default: 1000)}

\item{n.specific}{integer; number of top genes with high dataset-specificities. (default: 1000)}

\item{th.rank}{double; threshold for rank. (default: 0.1)}
}
\value{
a data.table object
}
\description{
highly variable genes finding (method using percetile rank values from F test)
}
\details{
For specific genes, the top n.specific genes with percetile rank values < th.rank are selected.
}
