% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.utils.R
\name{inSilico.TCell}
\alias{inSilico.TCell}
\title{Identification of T cells (simple average-threshold method)}
\usage{
inSilico.TCell(
  sce,
  out.prefix,
  assay.name = "norm_exprs",
  vis.v = c(0.25, 0.5),
  Th.CD3 = 0.25,
  Th.CD8 = 0.5,
  Th.CD4 = 0.5,
  Th.TH = 0.25,
  Th.TR = 0.25,
  do.zscore = F,
  Th.DC = 0.25,
  Th.GC1 = 0.25,
  Th.GC2 = 0.25,
  do.rescue = T
)
}
\arguments{
\item{sce}{object of \code{SingleCellExperiment}}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{assay.name}{character vector; which assay to use. (default: "norm_exprs").}

\item{vis.v}{double vector; for vertical lines in visulization. (default: c(0.25,0.5)).}

\item{Th.CD3}{double; threshold for T cell signature (mean of CD3D, CD3G). (default: 0.25)}

\item{Th.CD8}{double; threshold for CD8 (mean of CD8A, CD8B). (default: 0.25)}

\item{Th.CD4}{double; threshold for CD4 (expression of CD4). (default: 0.25)}

\item{Th.TH}{double; threshold for Thelper signature (mean of CD4, CD40LG). (default: 0.25)}

\item{Th.TR}{double; threshold for Treg signature (mean of CD4, FOXP3). (default: 0.25)}

\item{do.zscore}{logical; whether use zscore for calculation. (default: FALSE)}

\item{Th.DC}{double; threshold for delta receptor constant chain. (default: 0.25)}

\item{Th.GC1}{double; threshold for gamma receptor constant chain 1. (default: 0.25)}

\item{Th.GC2}{double; threshold for gamma receptor constant chain 2. (default: 0.25)}

\item{do.rescue}{logical; whether use "rescue" mode. (default: FALSE)}
}
\value{
a SingleCellExperiment object
}
\description{
Identification of T cells (simple average-threshold method)
}
\details{
columns stype and gdType will be added to the sce
}
