% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.utils.R
\name{inSilico.TGammaDelta}
\alias{inSilico.TGammaDelta}
\title{Identification of gamma delta T cells (simple average-threshold method)}
\usage{
inSilico.TGammaDelta(
  obj,
  out.prefix = NULL,
  assay.name = "norm_exprs",
  vis.v = c(0.25, 0.5),
  Th.CD3 = 0.25,
  Th.DC = 0.25,
  Th.GC1 = 0.25,
  Th.GC2 = 0.25
)
}
\arguments{
\item{obj}{object of \code{SingleCellExperiment}}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{assay.name}{character vector; which assay to use. (default: "norm_exprs").}

\item{vis.v}{double vector; for vertical lines in visulization. (default: c(0.25,0.5)).}

\item{Th.CD3}{double; threshold for T cell signature. (default: 0.25)}

\item{Th.DC}{double; threshold for delta receptor constant chain. (default: 0.25)}

\item{Th.GC1}{double; threshold for gamma receptor constant chain 1. (default: 0.25)}

\item{Th.GC2}{double; threshold for gamma receptor constant chain 2. (default: 0.25)}
}
\value{
a SingleCellExperiment object
}
\description{
Identification of gamma delta T cells (simple average-threshold method)
}
