% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{mergeDataFromFileTable}
\alias{mergeDataFromFileTable}
\title{merge data file file table}
\usage{
mergeDataFromFileTable(
  exp.list.table,
  gene.de.common,
  seu.list,
  sce.list,
  res.hi,
  method.clustering = "louvain",
  ncores = 12,
  cor.var = c("S.Score", "G2M.Score", "DIG.Score1"),
  contamination.vec = NULL,
  use.harmony = F
)
}
\arguments{
\item{exp.list.table}{data.table; one line for a dataset}

\item{gene.de.common}{character vector; common genes across multiple datasets}

\item{seu.list}{list; list containing seu objects}

\item{sce.list}{list; list containing sce objects}

\item{res.hi}{integer; high resolution used for mini-clusters identification}

\item{method.clustering}{character; clustering method for mini-clusters identification. (default: "louvain")}

\item{ncores}{integer; number of CPU cores to use. (default: 12)}

\item{cor.var}{character vector; subset of c("S.Score","G2M.Score","DIG.Score1","ISG.Score1","score.MALAT1")}

\item{contamination.vec}{character vector; cells to be excluded. (default: NULL)}

\item{use.harmony}{logical; use harmony to correct for batch effect (batches are defined in column batchV). (default: FALSE)}
}
\value{
a list containing 3 components: dat.avg, meta.extra.tb and clust
}
\description{
merge data file file table
}
\details{
For each dataset, the function first identify mini-clusters, then calculate the average expressions of mini-clusters.
}
