% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{mergeSCEDataFromFileTable}
\alias{mergeSCEDataFromFileTable}
\title{collapse gene by cell expression data to gene by mini-cluster expression data, with input stored in SingleCellExperiment object}
\usage{
mergeSCEDataFromFileTable(
  exp.list.table,
  gene.common,
  sce.list,
  group.vec,
  ncores = 6,
  contamination.vec = NULL
)
}
\arguments{
\item{exp.list.table}{data.table; one line for a dataset}

\item{gene.common}{character vector; common genes across multiple datasets}

\item{sce.list}{list; list containing sce objects}

\item{group.vec}{character; cell id to mini-cluster mapping vector}

\item{ncores}{integer; number of CPU cores to use}

\item{contamination.vec}{character vector; cells to be excluded. (default: NULL)}
}
\value{
a list, each component of which is expression matrix of gene.common by min-clusters
}
\description{
collapse gene by cell expression data to gene by mini-cluster expression data, with input stored in SingleCellExperiment object
}
\details{
For each dataset, the function first identify mini-clusters, then calculate the average expressions of mini-clusters.
}
