% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{render_KnitReport}
\alias{render_KnitReport}
\title{render a knitr report}
\usage{
render_KnitReport(template.file, out.file, par.list = list())
}
\arguments{
\item{template.file}{character; template file}

\item{out.file}{character; output file}

\item{par.list}{list; parameter list for rmarkdown::render}
}
\value{
a SingleCellExperiment object
}
\description{
render a knitr report
}
