% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{resetSig}
\alias{resetSig}
\title{reset the significance}
\usage{
resetSig(obj)
}
\arguments{
\item{obj}{object of \code{SingleCellExperiment} class}
}
\value{
a SingleCellExperiment object
}
\description{
reset the significance
}
