% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{run.inte.metaClust}
\alias{run.inte.metaClust}
\title{run the integration pipeline}
\usage{
run.inte.metaClust(
  exp.list.table,
  out.prefix,
  gene.exclude.file,
  nGene.common = 1500,
  nGene.specific = 0,
  ncores = 12,
  npc = 15,
  TH.gene.occ = 1,
  res.hi = 50,
  method.clustering = "louvain",
  cor.var = c("S.Score", "G2M.Score", "DIG.Score1"),
  contamination.vec = NULL
)
}
\arguments{
\item{exp.list.table}{data.table; one line for a dataset}

\item{out.prefix}{character; output prefix}

\item{gene.exclude.file}{character; file contains the genes to be excluded}

\item{nGene.common}{integer; number of common genes. (default: 1500)}

\item{nGene.specific}{integer; number of dataset specific genes. (default: 0)}

\item{ncores}{integer; number of CPU cores to use. (default: 12)}

\item{npc}{integer; number of principal components to use. (default: 15)}

\item{TH.gene.occ}{double; range from 0 to 1. genes present in >= TH.gene.occ datasets are used. (default: 1)}

\item{res.hi}{integer; high resolution used for mini-clusters identification. (default: 50)}

\item{method.clustering}{character; clustering method for mini-clusters identification. (default: "louvain")}

\item{cor.var}{character vector; subset of c("S.Score","G2M.Score","DIG.Score1","ISG.Score1","score.MALAT1")}

\item{contamination.vec}{character vector; cells to be excluded. (default: NULL)}
}
\value{
a list containing 3 components: sce.merged, seu.merged and meta.tb
}
\description{
run the integration pipeline
}
\details{
For each dataset, the function first identify mini-clusters, then calculate the average expressions of mini-clusters. The gene by mini-cluster expression data will pass the pipeline: PCA, harmony, UMAP/Clustering.
}
