% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{run.Leiden}
\alias{run.Leiden}
\title{wrapper for running leiden clustering}
\usage{
run.Leiden(dat.pca, SNN.k = 20, myseed = 123456, ...)
}
\arguments{
\item{dat.pca}{data.frame; row for sample, column for variable}

\item{SNN.k}{integer; parameter k of scran::buildSNNGraph. (default: 20)}

\item{myseed}{integer; seed for random number generation. (default: 123456)}

\item{...}{; other parameters passed to leiden::leiden}
}
\value{
a object returned from leiden::leiden
}
\description{
wrapper for running leiden clustering
}
\details{
implementation in Seurat (FindClusters) can only detect singletons :-(
}
