% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.utils.R
\name{run.Scanorama}
\alias{run.Scanorama}
\title{Wraper for running Scanorama}
\usage{
run.Scanorama(seu, col.batch = "batchV", assay.slot = "data", ...)
}
\arguments{
\item{seu}{object of \code{Seurat}}

\item{col.batch}{character; column name indicating batches (default: "batchV")}

\item{assay.slot}{character; which slot to be used? (default: "data")}

\item{...}{; passed to scanorama$integrate}
}
\value{
a Seurat object
}
\description{
Wraper for running Scanorama
}
\details{
run the Scanorama
}
