% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.utils.R
\name{run.scanpy}
\alias{run.scanpy}
\title{Wraper for running scanpy pipeline}
\usage{
run.scanpy(
  adata,
  out.prefix,
  gene.exclude.df,
  n.top = 1500,
  opt.res = "1",
  aid = "PRJ",
  plot.rd = c("umap"),
  opt.npc = 15,
  ncores = 16,
  res.test = seq(0.1, 2.4, 0.1),
  cor.var = c("S_score", "G2M_score", "DIG.Score", "percent.mito"),
  ncell.deg = 1500,
  do.deg = F,
  hvg.batch.minNCells = 100,
  use.harmony = T,
  method.integration = NULL,
  specie = "human",
  gene.mapping.table = NULL,
  res.addition = NULL,
  run.stage = 100
)
}
\arguments{
\item{adata}{object of \code{AnnData}}

\item{out.prefix}{character; output prefix}

\item{gene.exclude.df}{data.frame; gene blak list. Required column: seu.id.}

\item{n.top}{integer; number of top genes. (default: 1500)}

\item{opt.res}{character; optimal resolution (default: "1")}

\item{aid}{character; an ID (default: "PRJ")}

\item{plot.rd}{character vector; reducedDimNames used for plots (default: c("umap"))}

\item{opt.npc}{integer; optimal number of principal componets to use (default: 15)}

\item{ncores}{integer; number of CPU cores to use (default: 16)}

\item{res.test}{double; resolutions to test (default: seq(0.1,2.4,0.1) )}

\item{cor.var}{character vector; Subset of c("S_score","G2M_score","DIG.Score","ISG.Score","percent.mito") or NULL. (default: c("S_score","G2M_score","DIG.Score","percent.mito")).}

\item{ncell.deg}{integer; number of cell to downsample. used in the differentially expressed gene analysis. (default: 1500)}

\item{do.deg}{logical; whether perform the differentially expressed gene analysis. (default: FALSE)}

\item{hvg.batch.minNCells}{integer; required minimum number of cells in each batch. Batch with < hvg.batch.minNCells will not be used for HVG finding. (default: 100)}

\item{use.harmony}{logical; whether use the harmony method. (default: TRUE)}

\item{method.integration}{character; integration method. (default: NULL)}

\item{specie}{character; specie, one of "human", "mouse". (default: "human")}

\item{gene.mapping.table}{data.table; used for gene ID conversion. (default: NULL)}

\item{res.addition}{character vector; additional resolution parameters. Internally, resolutions from 0.1 to 2.4 will be used. (default: NULL)}

\item{run.stage}{integer; running stage. (default: 100)}
}
\value{
a list contain an object of \code{AnnData}
}
\description{
Wraper for running scanpy pipeline
}
\details{
run the scanpy pipeline
}
