% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.utils.R
\name{run.Seurat3}
\alias{run.Seurat3}
\title{Wraper for running Seurat3 pipeline}
\usage{
run.Seurat3(
  seu,
  sce,
  out.prefix,
  gene.exclude.df,
  n.top = 1500,
  measurement = "counts",
  platform = "10X",
  opt.res = "1",
  use.sctransform = F,
  aid = "PRJ",
  plot.rd = c("umap"),
  opt.npc = 15,
  ncores = 16,
  cor.var = c("S.Score", "G2M.Score", "DIG.Score1", "percent.mito", "batchV"),
  ncell.deg = 1500,
  do.deg = F,
  do.scale = F,
  use.harmony = F,
  method.integration = NULL,
  specie = "human",
  gene.mapping.table = NULL,
  res.addition = NULL,
  run.stage = 100
)
}
\arguments{
\item{seu}{object of \code{Seurat}}

\item{sce}{object of \code{SingleCellExperiment}}

\item{out.prefix}{character; output prefix}

\item{gene.exclude.df}{data.frame; gene blak list. Required column: seu.id.}

\item{n.top}{integer; number of top genes. (default: 1500)}

\item{measurement}{character; "counts", "TPM" or "cpm". (default: "counts")}

\item{platform}{character; "10X", "SmartSeq2", "InDrop" etc.. (default: "10X")}

\item{opt.res}{character; optimal resolution (default: "1")}

\item{use.sctransform}{logical; whether use scTransform method (default: FALSE)}

\item{aid}{character; an ID (default: "PRJ")}

\item{plot.rd}{character vector; reducedDimNames used for plots (default: c("umap"))}

\item{opt.npc}{integer; optimal number of principal componets to use (default: 15)}

\item{ncores}{integer; number of CPU cores to use (default: 16)}

\item{cor.var}{character vector; Subset of c("S.Score","G2M.Score","DIG.Score1","ISG.Score1","percent.mito","batchV") or NULL. (default: c("S.Score","G2M.Score","DIG.Score1","percent.mito","batchV")).}

\item{ncell.deg}{integer; number of cell to downsample. used in the differentially expressed gene analysis. (default: 1500)}

\item{do.deg}{logical; whether perform the differentially expressed gene analysis. (default: FALSE)}

\item{do.scale}{logical; whether scale the expression data. (default: FALSE)}

\item{use.harmony}{logical; whether use the harmony method. (default: FALSE)}

\item{method.integration}{character; integration method. (default: NULL)}

\item{specie}{character; specie, one of "human", "mouse". (default: "human")}

\item{gene.mapping.table}{data.table; used for gene ID conversion. (default: NULL)}

\item{res.addition}{character vector; additional resolution parameters. Internally, resolutions from 0.1 to 2.4 will be used. (default: NULL)}

\item{run.stage}{integer; running stage. (default: 100)}
}
\value{
a list contain a Seurat object and a SingleCellExperiment object
}
\description{
Wraper for running Seurat3 pipeline
}
\details{
run the Seurat3 pipeline
}
