% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.intClust.R
\name{sigGeneHeatmap}
\alias{sigGeneHeatmap}
\title{make heatmap for top signature genes}
\usage{
sigGeneHeatmap(
  out.prefix,
  gene.desc.top,
  sce.pb,
  gene.to.show.tb,
  value.var = "comb.ES",
  colSet = list(),
  ...
)
}
\arguments{
\item{out.prefix}{character; output prefix}

\item{gene.desc.top}{data.table; gene info table}

\item{sce.pb}{object of SingleCellExperiment; gene by meta-cluster values stored in this object}

\item{gene.to.show.tb}{data.table; infomation of genes to be shown}

\item{value.var}{character; stored value to use. (default: "comb.ES")}

\item{colSet}{list; color set to use. (default: list())}

\item{...}{; parameters passed to sscVis::plotMatrix.simple}
}
\description{
make heatmap for top signature genes
}
